wizard refactor: move wizardState -> lib/wizard
