+++ src/routes/wizard/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { page } from '$app/stores';
 	import { goto } from '$app/navigation';
-	import wizard from './wizardState';
+	import wizard from '$lib/wizard/store';
 	import { Section, WizardHeader, WizardNavItem } from '$lib/components';
 
 	if (!wizard?.initializing) {
