+++ src/routes/wizard/connect-wallet/balance/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import wizard from '../../wizardState';
+	import wizard from '$lib/wizard/store';
 	import { Button, WizardActions } from '$lib/components';
 	import WalletBalance from '$lib/wallet/WalletBalance.svelte';
 </script>
