+++ src/routes/wizard/connect-wallet/introduction/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import wizard from '../../wizardState';
+	import wizard from '$lib/wizard/store';
 	import { Button, WizardActions } from '$lib/components';
 
 	$wizard?.completed.add('introduction');
